requirements.txt
biopython
langchain-text-splitters
chromadb
google-generativeai
sentence-transformers